gps: change position accuracy to float
The accuracy returned from hybrid position can be very large, can
exceed range of uint_16. Thus the prototype of HEPE is changed to float.

+++ include/hardware/gps.h
@@ -646,7 +646,7 @@ typedef struct  {
     /** longitude in degrees */
     double longitude;
     /** Horzizontal error estimate in meters */
-    uint16_t HEPE;
+    float HEPE;
 } UlpNetworkPosition;
 
 /** Represents access point information */
